Chore/restructure (#102)
* restructure project
