+++ README.md
@@ -73,7 +73,7 @@ WisecondorX convert input.bam/cram output.npz [--optional arguments]
 | `--binsize x`                  | Size per bin in bp; the reference bin size should be a multiple of this value. Note that this parameter does not impact the resolution, yet it can be used to optimize processing speed (default: x=5e3) |
 | `--normdup`                    | Use this flag to avoid duplicate removal                                                                                                                                                                 |
 
-&rarr; Bash recipe at `./pipeline/convert.sh`
+&rarr; Bash recipe at `docs/include/pipeline/convert.sh`
 
 ### Stage (2) Create reference
 
@@ -91,7 +91,7 @@ WisecondorX newref reference_input_dir/*.npz reference_output.npz [--optional ar
 | `--plotyfrac x`                | plots Y read fraction histogram and Gaussian mixture fit to file x, can help when setting `--yfrac` manually; software quits after plotting |
 | `--cpus x`                     | Number of threads requested (default: x=1)                                                                                                  |
 
-&rarr; Bash recipe at `./pipeline/newref.sh`
+&rarr; Bash recipe at `docs/include/pipeline/newref.sh`
 
 ### Stage (3) Predict copy number alterations
 
@@ -116,7 +116,7 @@ WisecondorX predict test_input.npz reference_input.npz output_id [--optional arg
 
 <sup>**(\*)** At least one of these output formats should be selected</sup>
 
-&rarr; Bash recipe at `./pipeline/predict.sh`
+&rarr; Bash recipe at `docs/include/pipeline/predict.sh`
 
 ### Additional functionality
