+++ docs/include/example.plot/chr22.png
