+++ docs/include/example.plot/chr6.png
