+++ docs/include/pipeline/newref.sh
