+++ docs/include/pipeline/predict.sh
