+++ pyproject.toml
@@ -1,0 +1,17 @@
+[tool.black]
+line-length = 79
+target_version = ['py36','py37','py38','py39']
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+)/
+'''
