+++ setup.cfg
@@ -1,0 +1,63 @@
+[metadata]
+name = wisecondorx
+version = 1.2.6dev
+description = description="WisecondorX -- an evolved WISECONDOR"
+long_description = file: README.md
+author = "Matthias De Smet, Lennart Raman",
+author_email = "ict@cmgg.be",
+license = "by-nc-sa"
+license_file = LICENSE.md
+platforms = unix, linux, osx, win32
+keywords=
+        bioinformatics
+        biology
+        sequencing
+        NGS
+        next generation sequencing
+        CNV
+        SWGS
+        Shallow Whole Genome Sequencing
+classifiers =
+    Environment :: Console
+    Intended Audience :: Science/Research
+    Natural Language :: English
+    Operating System :: MacOS :: MacOS X
+    Operating System :: POSIX
+    Operating System :: Unix
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Topic :: Scientific/Engineering
+    Topic :: Scientific/Engineering :: Bio-Informatics
+
+
+[options]
+packages = wisecondorx
+package_dir =
+    =src
+include_package_data=True
+python_requires= >=3.6
+install_requires =
+    scipy
+    scikit-learn
+    pysam
+    numpy
+zip_safe = False
+
+[options.entry_points]
+    console_scripts =
+        wisecondorx = wisecondorx.main:main
+        WisecondorX = wisecondorx.main:main
+        wisecondorX = wisecondorx.main:main
+
+
+[options.extras_require]
+testing =
+    flake8
+    black
+    pre-commit
+
+[flake8]
+ignore = E266, E203, E501, C901, W503
+max-line-length = 79
+max-complexity = 18
