+++ src/wisecondorx/predict_output.py
@@ -4,7 +4,7 @@ import os
 
 import numpy as np
 
-from wisecondorX.overall_tools import (
+from wisecondorx.overall_tools import (
     exec_R,
     get_z_score,
     get_median_segment_variance,
