+++ src/wisecondorx/predict_tools.py
@@ -6,7 +6,7 @@ import numpy as np
 from scipy.stats import norm
 from sklearn.decomposition import PCA
 
-from wisecondorX.overall_tools import exec_R, get_z_score
+from wisecondorx.overall_tools import exec_R, get_z_score
 
 """
 Returns gender based on Gaussian mixture
